+++ src/json-file.js
@@ -1,5 +1,8 @@
 const fs = require('fs');
 
+// TODO: separate the concept of TranslationFile & DataTransform
+// JsonFile and YamlFile are completely the same, except for the
+// JSON.parse and JSON.stringify calls
 class JsonFile {
   // If a file is optional, it means it is okay for it to fail to load.
   constructor(filepath, optional) {
@@ -25,6 +28,12 @@ class JsonFile {
   async write() {
     let newFile = JSON.stringify(this.data, null, 2);
     await fs.promises.writeFile(this.filepath, newFile);
+  }
+
+  // For when you need to wholesale set new data, instead of mutating the
+  // existing data.
+  setData(newData) {
+    this.data = newData;
   }
 }